requirements.txt
Pillow~=8.4.0
imagehash~=4.2.1
imutils
opencv-python
Numpy
tqdm
scipy
numpy
piexif~=1.1.3
ocrd-pyexiftool
ZODB~=5.6.0
persistent~=4.7.0
transaction~=3.0.1
BTrees~=4.9.2
